docs: shuffle sections into logical order
The first section explains how to clone the repository, the second how
to build Infinitime with the docker image, but the details on actually
provisioning the image are at the end, despite this step taking place
before the build itself.

Move the sections into the order in which the steps should be followed.